osutil: make fault injection a build time feature
